requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.0.12
Django==3.0.9
django-bootstrap4==2.2.0
idna==3.3
isodate==0.6.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
rdflib==6.1.1
rdflib-jsonld==0.6.2
requests==2.28.0
simplejson==3.17.6
six==1.16.0
smartapi-python==1.2.9
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
websocket-client==1.3.3
